Split buffs mono-module into separate files
