+++ src/buffs/isProcEffect.ts
@@ -1,0 +1,14 @@
+/**
+ * @description Determine if a given effect object is a proc effect based on existing properties.
+ * Do note that it does not check the validity of each property, only the existence.
+ * @param effect object to check
+ * @returns whether the given effect object is considered a proc effect based on its properties
+ */
+export default function isProcEffect (effect: {
+	'proc id'?: string;
+	'unknown proc id'?: string;
+}): boolean {
+	return !!effect &&
+		typeof effect === 'object' &&
+		(Object.hasOwnProperty.call(effect, 'proc id') || Object.hasOwnProperty.call(effect, 'unknown proc id'));
+}
